requirements.txt
sanic==21.3.4
tortoise-orm[asyncpg]==0.16.21
sanic-openapi==21.3.2
aioredis
pydantic
aerich
webargs-sanic==2.0.0
marshmallow==3.12.1
